refactor : AcceptanceTest에서 중복된 코드가 아닌 것들을 각자 테스트 클래스로 이동
