+++ src/test/java/wooteco/subway/admin/acceptance/LineStationAcceptanceTest.java
@@ -2,11 +2,19 @@ package wooteco.subway.admin.acceptance;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
 import wooteco.subway.admin.dto.LineDetailResponse;
 import wooteco.subway.admin.dto.LineResponse;
 import wooteco.subway.admin.dto.StationResponse;
 
 import static org.assertj.core.api.Assertions.assertThat;
+
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+import java.util.HashMap;
+import java.util.Map;
 
 public class LineStationAcceptanceTest extends AcceptanceTest {
 
@@ -30,5 +38,16 @@ public class LineStationAcceptanceTest extends AcceptanceTest {
 
         LineDetailResponse lineResponseAfterRemoveLineStation = getLine(line2.getId());
         assertThat(lineResponseAfterRemoveLineStation.getStations().size()).isEqualTo(2);
+    }
+
+    void removeLineStation(Long lineId, Long stationId) {
+        given().
+            contentType(MediaType.APPLICATION_JSON_VALUE).
+            accept(MediaType.APPLICATION_JSON_VALUE).
+            when().
+            delete("/api/lines/" + lineId + "/stations/" + stationId).
+            then().
+            log().all().
+            statusCode(HttpStatus.NO_CONTENT.value());
     }
 }
