+++ src/test/java/wooteco/subway/admin/acceptance/StationAcceptanceTest.java
@@ -26,4 +26,21 @@ public class StationAcceptanceTest extends AcceptanceTest {
         List<StationResponse> stationsAfterDelete = getStations();
         assertThat(stationsAfterDelete.size()).isEqualTo(2);
     }
+
+    List<StationResponse> getStations() {
+        return
+            given().when().
+                get("/api/stations").
+                then().
+                log().all().
+                extract().
+                jsonPath().getList(".", StationResponse.class);
+    }
+
+    void deleteStation(Long id) {
+        given().when().
+            delete("/api/stations/" + id).
+            then().
+            log().all();
+    }
 }
